feat: fetch Slack users for changelog actors (#16)
This PR is part of our attempt to automate releases to staging. See the https://github.com/apify/apify-core/issues/17156 for full context and next steps.

**It introduces following changes:**

- New **`github-changelog-authors`** output, that will be populated with a JSON string of changelog commit authors. This will only work if the changelog mode is either `pull_request_commits`, or `commits_compare`.
- New **`fetch-author-slack-ids`** input,  that when set to `true` will use Slack API to enhance each returned author with a matching Slack ID.

Fetching Slack IDs is for now placed outside the critical path: If it fails, the action will still succeed. This should allow for graceful degradation, but if it works reliably, we can change it.

**Sample output:** 

```
[{name:"Tobiáš Potoček",email:"tobias.potocek@apify.com",slackId:"U0xyz"}]
```

**The PR also  tries to improve the overall DX:**

- There is a new `watch` script that automatically builds the code.
- There is a new `test_action` workflow that executes the action when pushed to a PR.
- The `README.md` has been significantly extended.